requirements.txt
Django>=5.1.4
djangorestframework>=3.14.0
Pillow>=10.1.0
